+++ src/modify/SecretLayer4.cpp
@@ -1,15 +1,15 @@
 #include "../SwelvyBG.hpp"
 #include <Geode/Geode.hpp>
-#include <Geode/modify/SecretLayer2.hpp>
+#include <Geode/modify/SecretLayer4.hpp>
 
 using namespace geode::prelude;
 
-class $modify(MySecretLayer2, SecretLayer2) {
+class $modify(MySecretLayer4, SecretLayer4) {
 	bool init() {
-		if (!SecretLayer2::init()) {
+		if (!SecretLayer4::init()) {
 			return false;
 		}
-		if (Mod::get()->getSettingValue<bool>("show-vault-of-secrets")){
+		if (Mod::get()->getSettingValue<bool>("show-chamber-of-time")){
 			if (auto bg = this->getChildByID("background")) {
 				bg->setVisible(false);
 				auto swelvyBG = SwelvyBG::create();